Moves DataProvider related methods into separated Trait
