--- src/Traits/WithAjaxDataProvider.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace YProjects\Forms\Traits;
-
-trait WithAjaxDataProvider
-{
-
-    protected bool $isAjax = false;
-
-    public function ajax(): self
-    {
-        $this->isAjax = true;
-        return $this;
-    }
-
-}
